Add custom exception, and move non-ecs related functions to __init__
